CI dosyalarına tensorflow paketleri eklendi
