+++ .github/workflows/ci.yml
@@ -96,7 +96,7 @@ jobs:
       - name: Install tensorflow(2.9.1) and tensorflow_hub(0.12.0)
         run: >
           pip install  tensorflow==2.9.1
-          pip install tensorflow_hub==0.12.0    
+          pip install tensorflow_hub==0.12.0  
 
       - name: Unittest for SAHI+YOLOV5/MMDET/Detectron2 on all platforms
         run: |
