+++ .github/workflows/package_testing.yml
@@ -94,7 +94,7 @@ jobs:
       - name: Install tensorflow(2.9.1) and tensorflow_hub(0.12.0)
         run: >
           pip install  tensorflow==2.9.1
-          pip install tensorflow_hub==0.12.0
+          pip install tensorflow_hub==0.12.0  
 
       - name: Install latest SAHI package
         run: >
